Fix h1, h2 headers on Wizard
