+++ server/db/migration/V20260517.1733__make_PARK_SPACES_optional_2856.sql
@@ -1,0 +1,7 @@
+/*
+  Deliverately don't specify a calcuationId, so this retroacively
+  makes PARK_SPACES optional for all existing progoram guidelines versions.
+*/
+
+UPDATE CalculationRule set required = 0
+WHERE code = 'PARK_SPACES'
